requirements.txt
albumentations==1.1.0
asynctest==0.13.0
cityscapesscripts==2.2.0
imagecorruptions==1.1.2
instaboostfast==0.1.2
kwarray==0.5.21
matplotlib==3.5.1
mmcv==1.4.0
mmcv_full==1.4.0
mmtrack==0.8.0
model_archiver==1.0.3
numpy==1.21.2
onnx==1.10.2
onnxruntime==1.10.0
opencv_python==4.5.4.60
Pillow==8.4.0
pycocotools==2.0.3
pytest==6.2.5
pytorch_sphinx_theme==0.0.19
PyYAML==6.0
requests==2.26.0
scikit_learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools==58.0.4
six==1.16.0
tensorrt==0.0.1
terminaltables==3.1.10
torch==1.7.0
ts==0.5.1
xlrd==2.0.1
xlutils==2.0.0
